petweb-api/requirements-win10-py36-rtx2080ti.txt
absl-py==0.13.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
dipy==1.3.0
elasticdeform==0.4.9
gast==0.5.2
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.6.3
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
nibabel==3.1.0
nilearn==0.7.0
numpy==1.19.5
opencv-python==4.3.0.36
packaging==21.0
pandas==1.1.5
pathtools==0.1.2
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboard==1.14.0
tensorflow-gpu==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
watchdog==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
